update-deployment: handle empty username in secret
When the user edits the secret from the OpenShift console, OpenShift
adds an empty `username` field. Afterwards, the mounted secret *will*
contain a username file but it will be empty. This causes the
update-deployment task to use https://:password@repo_url as the origin
url, which makes git authentication fail.

Fix by checking not just the file existence, but also that the username
is non-empty.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>