Push EC bundle data to konflux-ci/tekton-catalog
STONEBLD-2339

Push the data-acceptable-bundles image to konflux-ci/tekton-catalog.

And DO NOT stop pushing it to redhat-appstudio-tekton-catalog. All the
existing EnterpriseContractPolicies reference
redhat-appstudio-tekton-catalog. If we stopped pushing the
data-acceptable-bundles image there, the acceptable bundles data would
effectively stop getting updated until we manage to update every ECP.

Update the ECP in this repo to use the new location of the bundle image.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>